* enc/unicode.c: fix implicit conversion error with clang. fixup r53548. * string.c: ditto.
git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@53552 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ enc/unicode.c
@@ -630,5 +630,5 @@ onigenc_unicode_case_map(OnigCaseFoldType* flags,
 	}
 	to += ONIGENC_CODE_TO_MBC(enc, code, to);
     }
-    return to-to_start;
+    return (int)(to-to_start);
 }
